added mysql migration:
I introduced two new flag that are governing the migration process. For a detailed description of what the new flags do see the help promt from the binary.

Its save to always keep the flag for debugging purposes but be careful with production databases as you are at the mercy of whats inside of the migration folder.

The migration is based on the tool that was previosly used and is now included in the binary.